package_requirements.txt
azureml-core==1.18.0
gensim==3.8.3
matplotlib==3.3.3
numpy==1.19.1
requests==2.24.0
scikit-learn==0.23.2
seaborn==0.11.1
TakeSentenceTokenizer==1.0.1
torch==1.7.1